Remove ./lib/make-error.js file and replace all the `makeError` invocations with `new MulterError` class
