+++ index.js
@@ -1,4 +1,3 @@
-var makeError = require('./lib/make-error')
 var makeMiddleware = require('./lib/make-middleware')
 
 var diskStorage = require('./storage/disk')
@@ -38,7 +37,7 @@ Multer.prototype._makeMiddleware = function (fields, fileStrategy) {
 
     function wrappedFileFilter (req, file, cb) {
       if ((filesLeft[file.fieldname] || 0) <= 0) {
-        return cb(makeError('LIMIT_UNEXPECTED_FILE', file.fieldname))
+        return cb(new MulterError('LIMIT_UNEXPECTED_FILE', file.fieldname))
       }
 
       filesLeft[file.fieldname] -= 1
