+++ lib/make-middleware.js
@@ -5,7 +5,7 @@ var onFinished = require('on-finished')
 var appendField = require('append-field')
 
 var Counter = require('./counter')
-var makeError = require('./make-error')
+var MulterError = require('./multer-error')
 var FileAppender = require('./file-appender')
 var removeUploadedFiles = require('./remove-uploaded-files')
 
@@ -76,7 +76,7 @@ function makeMiddleware (setup) {
     }
 
     function abortWithCode (code, optionalField) {
-      abortWithError(makeError(code, optionalField))
+      abortWithError(new MulterError(code, optionalField))
     }
 
     // handle text field data
