+++ lib/multer-error.js
@@ -1,9 +1,19 @@
 var util = require('util')
 
-function MulterError (message, code, field) {
+var errorMessages = {
+  'LIMIT_PART_COUNT': 'Too many parts',
+  'LIMIT_FILE_SIZE': 'File too large',
+  'LIMIT_FILE_COUNT': 'Too many files',
+  'LIMIT_FIELD_KEY': 'Field name too long',
+  'LIMIT_FIELD_VALUE': 'Field value too long',
+  'LIMIT_FIELD_COUNT': 'Too many fields',
+  'LIMIT_UNEXPECTED_FILE': 'Unexpected field'
+}
+
+function MulterError (code, field) {
   Error.captureStackTrace(this, this.constructor)
   this.name = this.constructor.name
-  this.message = message
+  this.message = errorMessages[code]
   this.code = code
   if (field) this.field = field
 }
